+++ test/init.ts
@@ -35,3 +35,7 @@ if (opts['preact-lib']) {
 // Log details of which Preact library is being used.
 import { isPreact10 } from '../src/util';
 console.log(`Using Preact ${isPreact10() ? '10+' : '<= 9'}`);
+
+// For Preact <= 9, modify VNode class for compatibility with Preact 10.
+import { addTypeAndPropsToVNode } from '../src/compat';
+addTypeAndPropsToVNode();