+++ test/shallow-render-utils-test.tsx
@@ -1,8 +1,13 @@
-import { Fragment, cloneElement, h } from 'preact';
+import { Component, Fragment, VNode, cloneElement, h } from 'preact';
 import { assert } from 'chai';
 
-import { getRealType, withShallowRendering } from '../src/shallow-render-utils';
-import { componentForDOMNode, render } from '../src/compat';
+import {
+  getRealType,
+  withShallowRendering,
+  isShallowRendered,
+  shallowRenderVNodeTree,
+} from '../src/shallow-render-utils';
+import { componentForDOMNode, render, childElements } from '../src/compat';
 import { isPreact10 } from '../src/util';
 
 describe('shallow-render-utils', () => {
@@ -31,8 +36,13 @@ describe('shallow-render-utils', () => {
   describe('withShallowRendering', () => {
     it('replaces child components with placeholders', () => {
       const fullOutput = 'Hello<span>world</span>';
-      const shallowOutput =
-        'Hello<shallow-render component="Child"></shallow-render>';
+      let shallowOutput: string;
+      if (isPreact10()) {
+        shallowOutput = 'Hello';
+      } else {
+        shallowOutput =
+          'Hello<shallow-render component="Child"></shallow-render>';
+      }
 
       // Normal render should return full output.
       const el = <Component />;
@@ -89,11 +99,43 @@ describe('shallow-render-utils', () => {
       withShallowRendering(() => {
         render(el, container);
       });
-      const child = container.querySelector('shallow-render')!;
-      const childComponent = componentForDOMNode(child)!;
+      let childComponent: Component;
+      if (isPreact10()) {
+        const fragVNode = (container as any)._prevVNode;
+        const rootComponent = fragVNode._children[0]._component;
+        childComponent = rootComponent._prevVNode._children[1]._component;
+      } else {
+        const child = container.querySelector('shallow-render')!;
+        childComponent = componentForDOMNode(child)!;
+      }
       assert.ok(childComponent);
       assert.equal(childComponent instanceof Child, false);
       assert.equal(getRealType(childComponent), Child);
+    });
+  });
+
+  describe('shallowRenderVNodeTree', () => {
+    it('modifies nodes to shallow-render', () => {
+      function Parent() {
+        return null;
+      }
+      function Child() {
+        return null;
+      }
+      const el = (
+        <Parent>
+          <Child />
+        </Parent>
+      );
+      const childEl = childElements(el)[0] as VNode;
+
+      assert.isFalse(isShallowRendered(el));
+      assert.isFalse(isShallowRendered(childEl));
+
+      shallowRenderVNodeTree(el);
+
+      assert.isTrue(isShallowRendered(el));
+      assert.isTrue(isShallowRendered(childEl));
     });
   });
 });